Added application support
Moved code to proper package. Added main class for the application build.
