+++ build.gradle
@@ -21,4 +21,10 @@ javafx {
     modules = [ 'javafx.controls',  'javafx.fxml', 'javafx.base' ]
 }
 
-mainClassName = 'Main'
+jar {
+    manifest {
+        attributes 'Main-Class': 'com.tcb12.application.Main'
+    }
+}
+
+mainClassName = 'com.tcb12.application.Main'
